+++ back/src/bsffs/resolvers/mutations/__tests__/createDraftBsff.integration.ts
@@ -1,4 +1,9 @@
-import { UserRole, BsffStatus, BsffType } from "@prisma/client";
+import {
+  UserRole,
+  BsffStatus,
+  BsffType,
+  BsffPackagingType
+} from "@prisma/client";
 import { resetDatabase } from "../../../../../integration-tests/helper";
 import {
   Mutation,
@@ -29,6 +34,9 @@ const CREATE_DRAFT_BSFF = `
       }
       packagings {
         id
+        name
+        type
+        other
       }
     }
   }
@@ -334,7 +342,12 @@ describe("Mutation.createDraftBsff", () => {
               }
             },
             packagings: [
-              { name: "Bouteille", volume: 1, weight: 1, numero: "cont1" }
+              {
+                type: BsffPackagingType.BOUTEILLE,
+                volume: 1,
+                weight: 1,
+                numero: "cont1"
+              }
             ],
             repackaging: previousBsffs.flatMap(previousBsff =>
               previousBsff.packagings.map(p => p.id)
@@ -533,5 +546,118 @@ describe("Mutation.createDraftBsff", () => {
         ]);
       }
     );
+
+    it("should allow to create a BSFF packaging with deprecated field `name=bouteille`", async () => {
+      const emitter = await userWithCompanyFactory("ADMIN");
+
+      const { mutate } = makeClient(emitter.user);
+      const { data, errors } = await mutate<
+        Pick<Mutation, "createDraftBsff">,
+        MutationCreateDraftBsffArgs
+      >(CREATE_DRAFT_BSFF, {
+        variables: {
+          input: {
+            type: BsffType.COLLECTE_PETITES_QUANTITES,
+            emitter: {
+              company: {
+                name: emitter.company.name,
+                siret: emitter.company.siret,
+                address: emitter.company.address,
+                contact: emitter.user.name,
+                mail: emitter.user.email
+              }
+            },
+            packagings: [
+              { name: "Bouteille", numero: "cont1", volume: 1, weight: 1 }
+            ]
+          }
+        }
+      });
+      expect(errors).toBeUndefined();
+      expect(data.createDraftBsff.packagings[0].name).toEqual("BOUTEILLE");
+      expect(data.createDraftBsff.packagings[0].type).toEqual("BOUTEILLE");
+      expect(data.createDraftBsff.packagings[0].other).toBeNull();
+    });
+
+    it("should allow to create a BSFF packaging with deprecated field `name=anything`", async () => {
+      const emitter = await userWithCompanyFactory("ADMIN");
+
+      const { mutate } = makeClient(emitter.user);
+      const { data, errors } = await mutate<
+        Pick<Mutation, "createDraftBsff">,
+        MutationCreateDraftBsffArgs
+      >(CREATE_DRAFT_BSFF, {
+        variables: {
+          input: {
+            type: BsffType.COLLECTE_PETITES_QUANTITES,
+            emitter: {
+              company: {
+                name: emitter.company.name,
+                siret: emitter.company.siret,
+                address: emitter.company.address,
+                contact: emitter.user.name,
+                mail: emitter.user.email
+              }
+            },
+            packagings: [
+              {
+                name: "Bouteille de récup",
+                numero: "cont1",
+                volume: 1,
+                weight: 1
+              }
+            ]
+          }
+        }
+      });
+      expect(errors).toBeUndefined();
+      expect(data.createDraftBsff.packagings[0].type).toEqual("AUTRE");
+      expect(data.createDraftBsff.packagings[0].name).toEqual(
+        "Bouteille de récup"
+      );
+      expect(data.createDraftBsff.packagings[0].other).toEqual(
+        "Bouteille de récup"
+      );
+    });
+
+    it("should throw error when passing both name= and type= on a BSFF packagng", async () => {
+      const emitter = await userWithCompanyFactory("ADMIN");
+
+      const { mutate } = makeClient(emitter.user);
+      const { errors } = await mutate<
+        Pick<Mutation, "createDraftBsff">,
+        MutationCreateDraftBsffArgs
+      >(CREATE_DRAFT_BSFF, {
+        variables: {
+          input: {
+            type: BsffType.COLLECTE_PETITES_QUANTITES,
+            emitter: {
+              company: {
+                name: emitter.company.name,
+                siret: emitter.company.siret,
+                address: emitter.company.address,
+                contact: emitter.user.name,
+                mail: emitter.user.email
+              }
+            },
+            packagings: [
+              {
+                name: "Bouteille de récup",
+                type: "BOUTEILLE",
+                numero: "cont1",
+                volume: 1,
+                weight: 1
+              }
+            ]
+          }
+        }
+      });
+      expect(errors).toEqual([
+        expect.objectContaining({
+          message:
+            "Vous ne pouvez pas préciser à la fois le champ `type` et le champ `name`"
+        })
+      ]);
+    });
   });
 });