+++ back/src/bsffs/resolvers/mutations/__tests__/publishBsff.integration.ts
@@ -1,4 +1,4 @@
-import { UserRole } from "@prisma/client";
+import { BsffPackagingType, UserRole } from "@prisma/client";
 import { gql } from "apollo-server-core";
 import { resetDatabase } from "../../../../../integration-tests/helper";
 import getReadableId, { ReadableIdPrefix } from "../../../../forms/readableId";
@@ -106,7 +106,12 @@ describe("publishBsff", () => {
         weightValue: 1,
         weightIsEstimate: false,
         packagings: {
-          create: { name: "BOUTEILLE", numero: "123", weight: 1, volume: 1 }
+          create: {
+            type: BsffPackagingType.BOUTEILLE,
+            numero: "123",
+            weight: 1,
+            volume: 1
+          }
         }
       }
     });